Don't override occurrence indices with lower frequency entries
Frequency dictionaries often have multiple entries for the same
expression; usually an entry where the expression is written with kanji,
and an entry where the expression is written in kana. For example:

  ['指摘', 'freq', {'reading': 'してき', 'frequency': {'value': 3817, 'displayValue': '3817'}}]
  ['指摘', 'freq', {'reading': 'してき', 'frequency': {'value': 158844, 'displayValue': '158844㋕'}}]

In this case, when building the OccurenceIndex, the first entry with
frequency 3817 is added to the index. Then later, the second entry with
frequency 158844 is added to the index. Instead, entries with the
highest frequency should be prioritized.

This change fixes this issue by only overriding existing entries for an
expression if the new entry has a higher frequency.